chore: Only deploy web version on dispatch or merge with main (#17)
Updates the workflow so it it doesn't automatically deploy the web
version for every change to a PR, but instead when it's merged with
main.

Changes in export job conditions:

*
[`.github/workflows/export.yml`](diffhunk://#diff-abd067ac9ffcb98dd6b5da7a1bff2138d3a3a9ced0ba8f74886fbbe3d7686540L82-R82):
Updated the `if` conditions for the `export_windows` job to include push
events.
*
[`.github/workflows/export.yml`](diffhunk://#diff-abd067ac9ffcb98dd6b5da7a1bff2138d3a3a9ced0ba8f74886fbbe3d7686540L117-R117):
Updated the `if` conditions for the `export_linux` job to include push
events.
*
[`.github/workflows/export.yml`](diffhunk://#diff-abd067ac9ffcb98dd6b5da7a1bff2138d3a3a9ced0ba8f74886fbbe3d7686540L152-R152):
Updated the `if` conditions for the `export_web` job to include push
events.
*
[`.github/workflows/export.yml`](diffhunk://#diff-abd067ac9ffcb98dd6b5da7a1bff2138d3a3a9ced0ba8f74886fbbe3d7686540R176):
Added a condition to the `Deploy to GitHub Pages` step to include push
events.
*
[`.github/workflows/export.yml`](diffhunk://#diff-abd067ac9ffcb98dd6b5da7a1bff2138d3a3a9ced0ba8f74886fbbe3d7686540L193-R194):
Updated the `if` conditions for the `export_mac` job to include push
events.